requirements.txt
pyTelegramBotAPI==4.12.0
python-dotenv==0.21.0
schedule==1.2.0
